condition-lib/requirements.txt
gradio
psycopg2-binary
python-dotenv==1.0.1
